+++ audio/Square.cpp
@@ -23,7 +23,7 @@ Square::Square()
      
  }
 
-float Square::calculate(float fTime, float fFrequency)
+double Square::calculate(double fTime, double fFrequency)
 {
     return sin( 2. * M_PI * fFrequency * fTime) > 0 ? 1.0 : -1.0;
 }
